MAINT: Rename _path_utils to _utils
